♻️ : #38 - useModalClose hook 생성 & 적용
