+++ src/hooks/useModalClose.ts
@@ -1,0 +1,20 @@
+'use client';
+
+import { useEffect } from 'react';
+
+export const useModalClose = (
+  ref: React.RefObject<HTMLElement>,
+  callback: () => void,
+) => {
+  const handleClickOutside = (e: MouseEvent) => {
+    if (ref.current && !ref.current.contains(e.target as Node)) {
+      callback();
+    }
+  };
+  useEffect(() => {
+    document.addEventListener('mousedown', handleClickOutside);
+    return () => {
+      document.removeEventListener('mousedown', handleClickOutside);
+    };
+  });
+};
